addFriend & fix files
