--- controllers/api/index.js
@@ -1,7 +1,0 @@
-const express = require('express');
-const router = express.Router();
-
-const userRoutes = require('./userRoutes');
-router.use('/users', userRoutes);
-
-module.exports = router;
